d3d::Include refactoring
- drop `from_fn` (doesn't track CWD, kinda useless)
- `from_fn_with_header` -> `from_blob_meta_fn`
- parent meta passed only now

This restricts the default d3d::Include interface a bit.
In particular, scanning the includer source to do magic would require storing a copy of said source in metadata.
That's a rather awkward/evil bit of nonsense to try and support, and just makes the parent Option more awkward to use 99% of the time.
A future `from_blob_meta_src_fn` could be added later to revive that functionality without the extra copy if actually desired.
A seeded-meta version should perhaps be considered to get rid of the Option...